+++ web/src/components/landing/index.ts
@@ -1,0 +1,8 @@
+export { LandingHero } from "./LandingHero";
+export { DayTimeline } from "./DayTimeline";
+export { FeaturesBento } from "./FeaturesBento";
+export { StatsStrip } from "./StatsStrip";
+export { HowItWorksSteps } from "./HowItWorksSteps";
+export { Faq } from "./Faq";
+export { FinalCta } from "./FinalCta";
+export { LandingFooter } from "./LandingFooter";